combine actions e2e specs
